requirements.txt
Flask==2.2.2
Flask_Login==0.6.2
Flask_SQLAlchemy==2.5.1
mysql_connector_repackaged==0.3.1
SQLAlchemy==1.4.41
Werkzeug==2.2.2
